pyproject.toml
[project]
name = "sa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "boto3>=1.42.10",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]

deployment = [
    "boto3>=1.42.10",
    "joblib>=1.5.3",
    "onnxscript>=0.5.6",
    "sentence-transformers>=5.2.0",
    "torch>=2.8.0",
    "scikit-learn==1.7.2",
    "skl2onnx>=1.19.1",
    "transformers>=4.57.3",
]

inference = [
    "onnxruntime>=1.23.2",
    "tokenizers>=0.22.1",
]
